requirements.txt
pynest-api==0.3.1
asyncpg==0.29.0
uuid7==0.1.0
psycopg2_binary==2.9.10
pydantic==1.10.13
python-dotenv==1.0.0
sqlalchemy==2.0.19
alembic>=1.7.0
email-validator>=1.1.3
python-jose>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.5
structlog>=21.1.0
pyyaml>=6.0
python-jose[cryptography]>=3.3.0
